A few new responses
I figured we could add a couple of new responses for her. I added new responses pertaining to the "I need" statement, the "Why can't I?" statement, and the mother/father statements. If you don't want to put them in, that's okay. Thanks for taking the time to look it over though. 
~Ulvsblakk